+++ packages/expo-image/src/Image.types.ts
@@ -227,6 +227,13 @@ export type ImageProps = {
    * @platform ios
    */
   accessibilityLabel?: string | null;
+
+  /**
+   * Enables Live Text interaction with the image. Check official [Apple documentation](https://developer.apple.com/documentation/visionkit/enabling_live_text_interactions_with_images) for more details.
+   * @default false
+   * @platform ios 16.0+
+   */
+  enableLiveTextInteraction?: boolean;
 };
 
 /**